AE summary table

AE summary table

Usage

gt_ae_summary(
  outdata,
  source,
  analysis,
  title = c("analysis", "observation", "population"),
  footnotes = NULL
)

Arguments

outdata

An outdata object created by prepare_ae_specific().

source

A character value of the data source.

analysis

One of analysis name existing at outdata$meta$analysis

title

Term "analysis", "observation"and "population") for collecting title from metadata or a character vector of table titles.

footnotes

A character vector of table footnotes.

Value

RTF file and the source dataset for AE summary table.

Examples

library(gt)
#' # Define metadata
adsl <- forestly::forestly_adsl
adae <- forestly::forestly_adae

adsl$TRT01A <- factor(
  adsl$TRT01A,
  levels = c("Xanomeline Low Dose", "Placebo"),
  labels = c("Low Dose", "Placebo")
)
adae$TRTA <- factor(
  adae$TRTA,
  levels = c("Xanomeline Low Dose", "Placebo"),
  labels = c("Low Dose", "Placebo")
)

analysis_plan <- metalite::plan(
  analysis = "ae_summary",
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser"
)

meta <- metalite::meta_adam(observation = adae, population = adsl) |>
  metalite::define_plan(analysis_plan) |>
  metalite::define_population(
    name = "apat",
    var = c(
      "USUBJID", "SAFFL", "TRT01A", "TRTDUR",
      "SITEID", "SEX", "RACE", "AGE"
    ),
    group = "TRT01A",
    subset = SAFFL == "Y",
    label = "All Participants as Treated"
  ) |>
  metalite::define_observation(
    name = "wk12",
    var = c(
      "USUBJID", "SAFFL", "TRTA", "AEDECOD", "AEBODSYS", "AEREL",
      "AESER", "AEOUT", "AEACN", "AESDTH", "ASTDT", "AENDT"
    ),
    group = "TRTA",
    subset = SAFFL == "Y",
    label = "Weeks 0 to 12"
  ) |>
  metalite::define_parameter(
    name = "any",
    term1 = "",
    term2 = "",
    var = "AEDECOD",
    soc = "AEBODSYS",
    label = "All AEs"
  ) |>
  metalite::define_parameter(
    name = "rel",
    term1 = "Drug-Related",
    term2 = "",
    subset = AEREL %in% c("POSSIBLE", "PROBABLE"),
    var = "AEDECOD",
    soc = "AEBODSYS",
    label = "Drug-related AEs"
  ) |>
  metalite::define_parameter(
    name = "ser",
    term1 = "Serious",
    term2 = "",
    subset = AESER == "Y",
    var = "AEDECOD",
    soc = "AEBODSYS",
    label = "Serious AEs"
  ) |>
  metalite::define_analysis(
    name = "ae_summary",
    title = "Adverse Event Summary"
  ) |>
  metalite::meta_build()

outdata <- prepare_ae_summary(meta,
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser"
)
#> any
#> rel
#> ser
outdata |>
  format_ae_summary() |>
  gt_ae_summary(
    analysis = "ae_summary",
    source = "Source:  [CDISCpilot: adam-adsl; adae]"
  )
#> [1] "Adverse Event Summary"       "Weeks 0 to 12"              
#> [3] "All Participants as Treated"

  

Adverse Event Summary

Weeks 0 to 12

All Participants as Treated
    

      	
        
Low Dose

      
      	
        
Placebo

      
      	
        
Total

      
    

n
      	(%)
      	n
      	(%)
      	n
      	(%)
    

Participants in population
	84

	86

	170

with one or more adverse events
	77
	(91.7)
	69
	(80.2)
	146
	(85.9)

with no adverse events
	7
	 (8.3)
	17
	(19.8)
	24
	(14.1)

with drug-relateda adverse events
	73
	(86.9)
	44
	(51.2)
	117
	(68.8)

with serious adverse events
	1
	 (1.2)
	0
	 (0.0)
	1
	 (0.6)

aDetermined by the investigator to be related to the drug.
    

Source:  [CDISCpilot: adam-adsl; adae]