Format AE summary analysis

Format AE summary analysis

Usage

format_ae_summary(
  outdata,
  display = c("n", "prop", "total"),
  digits_prop = 1,
  digits_ci = 1,
  digits_p = 3,
  digits_dur = c(1, 1),
  digits_events = c(1, 1),
  mock = FALSE
)

Arguments

outdata

a outdata object created by prepare_ae_specific

display

a character vector of measurement to be displayed.

n: number of subjects with AE.

prop: proportion of subjects with AE.

total: total columns

diff: risk difference

diff_ci: 95% confidence interval of risk difference using M&N method

diff_p: p-value of risk difference using M&N method

dur: average of AE duration

events: average number of AE per subject

digits_prop

a numeric value of number of digits for proportion value.

digits_ci

a numeric value of number of digits for confidence interval

digits_p

a numeric value of number of digits for p-value .

digits_dur

a numeric value of number of digits for average duration of AE

digits_events

a numeric value of number of digits for average of number of AE per subjects.

mock

a boolean value to display mock table