Format AE summary analysis

Format AE summary analysis

Usage

format_ae_summary(
  outdata,
  display = c("n", "prop", "total"),
  digits_prop = 1,
  digits_ci = 1,
  digits_p = 3,
  digits_dur = c(1, 1),
  digits_events = c(1, 1),
  mock = FALSE
)

Arguments

outdata

A outdata object created by prepare_ae_specific().

display

A character vector of measurement to be displayed:

n: Number of subjects with AE.

prop: Proportion of subjects with AE.

total: Total columns.

diff: Risk difference.

diff_ci: 95% confidence interval of risk difference using M&N method.

diff_p: p-value of risk difference using M&N method.

dur: Average of AE duration.

events: Average number of AE per subject.

digits_prop

A numeric value of number of digits for proportion value.

digits_ci

A numeric value of number of digits for confidence interval.

digits_p

A numeric value of number of digits for p-value.

digits_dur

A numeric value of number of digits for average duration of AE.

digits_events

A numeric value of number of digits for average of number of AE per subjects.

mock

A boolean value to display mock table.

Value

A list of analysis raw datasets.

Examples

meta <- meta_ae_example()
outdata <- prepare_ae_summary(meta,
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser"
)
#> [1] "any"
#> [1] "rel"
#> [1] "ser"
tbl <- outdata |>
  format_ae_summary()
head(tbl$tbl)
#>                                    name n_1 prop_1 n_2 prop_2 n_3 prop_3 n_4
#> 1            Participants in population  86   <NA>  84   <NA>  84   <NA> 254
#> 2       with one or more adverse events  69 (80.2)  77 (91.7)  79 (94.0) 225
#> 3                with no adverse events  17 (19.8)   7  (8.3)   5  (6.0)  29
#> 21 with drug-related{^a} adverse events  44 (51.2)  73 (86.9)  70 (83.3) 187
#> 22          with serious adverse events   0  (0.0)   1  (1.2)   2  (2.4)   3
#>    prop_4
#> 1    <NA>
#> 2  (88.6)
#> 3  (11.4)
#> 21 (73.6)
#> 22  (1.2)