Specific Adverse Events Table

Specific Adverse Events Table

Usage

tlf_ae_specific(
  outdata,
  medra_version,
  source,
  col_rel_width = NULL,
  text_font_size = 9,
  orientation = "portrait",
  footnotes = NULL,
  path_outdata = NULL,
  path_outtable = NULL
)

Arguments

outdata

a outdata list created from prepare_ae_specific

medra_version

a character value of the MedDRA Version for this dataset

source

a character value of the data source

col_rel_width

Column relative width in a vector e.g. c(2,1,1) refers to 2:1:1. Default is NULL for equal column width.

text_font_size

Text font size. To vary text font size by column, use numeric vector with length of vector equal to number of columns displayed e.g. c(9,20,40).

orientation

Orientation in 'portrait' or 'landscape'.

footnotes

a character vector of table footnotes

path_outdata

a character string of the outdata path.

path_outtable

a character string of the outtable path.

Examples

library(r2rtf)
library(metalite)
meta <- meta_ae_dummy()

meta |>
  prepare_ae_specific(
    population = "apat",
    observation = "wk12",
    parameter = "rel"
  ) |>
  format_ae_specific() |>
  tlf_ae_specific(
    source = "Source:  [CDISCpilot: adam-adsl; adae]",
    medra_version = "24.0",
    path_outdata = tempfile(fileext = ".Rdata"),
    path_outtable = tempfile(fileext = ".rtf")
  )
#> The outdata is saved in/tmp/RtmpmdKMla/file3908701d70f7.Rdata
#> The output is saved in/tmp/RtmpmdKMla/file39086d973a20.rtf