AE summary table

AE summary table

Usage

tlf_ae_summary(
  outdata,
  source,
  col_rel_width = NULL,
  text_font_size = 9,
  orientation = "portrait",
  title = c("analysis", "observation", "population"),
  footnotes = NULL,
  path_outdata = NULL,
  path_outtable = NULL
)

Arguments

outdata

An outdata object created by prepare_ae_specific().

source

A character value of the data source.

col_rel_width

Column relative width in a vector e.g. c(2,1,1) refers to 2:1:1. Default is NULL for equal column width.

text_font_size

Text font size. To vary text font size by column, use numeric vector with length of vector equal to number of columns displayed e.g. c(9,20,40).

orientation

Orientation in 'portrait' or 'landscape'.

title

Term "analysis", "observation"and "population") for collecting title from metadata or a character vector of table titles.

footnotes

A character vector of table footnotes.

path_outdata

A character string of the outdata path.

path_outtable

A character string of the outtable path.

Value

RTF file and the source dataset for AE summary table.

Examples

meta <- meta_ae_example()
outdata <- prepare_ae_summary(meta,
  population = "apat",
  observation = "wk12",
  parameter = "any;rel;ser"
)
#> any
#> rel
#> ser
outdata |>
  format_ae_summary() |>
  tlf_ae_summary(
    source = "Source:  [CDISCpilot: adam-adsl; adae]",
    path_outdata = tempfile(fileext = ".Rdata"),
    path_outtable = tempfile(fileext = ".rtf")
  )
#> The outdata is saved in/tmp/RtmpTtOPk8/file179c47dd4e0e.Rdata
#> The output is saved in/tmp/RtmpTtOPk8/file179c7742152f.rtf